remove pyqtdarktheme

+++ src/presentation/user/dash/view.py
@@ -5,7 +5,8 @@ from PyQt5 import QtWidgets as qtw
 from loguru import logger
 
 from src import domain
-from src.presentation.shared import icons, widgets, font
+from src.presentation.shared import widgets
+from src.presentation.shared.theme import font, icons
 from src.presentation.shared.widgets import table_view, popup
 from src.presentation.user.dash import requests
 from src.presentation.user.dash.state import UserDashState
@@ -32,57 +33,68 @@ class UserDash(qtw.QWidget):
         self._refresh_btn.setFixedWidth(font.BOLD_FONT_METRICS.height() + 8)
         self._refresh_btn.setToolTip("Refresh")
 
-        add_btn_icon = icons.add_btn_icon(parent=self)
-        self._add_btn = qtw.QPushButton(add_btn_icon, "")
-        self._add_btn.setFixedWidth(font.BOLD_FONT_METRICS.height() + 8)
-        self._add_btn.setToolTip("Add New User")
-
         toolbar = qtw.QHBoxLayout()
         toolbar.addWidget(self._refresh_btn)
-        toolbar.addWidget(self._add_btn)
+
+        if self._current_user.is_admin:
+            add_btn_icon = icons.add_btn_icon(parent=self)
+            self._add_btn = qtw.QPushButton(add_btn_icon, "")
+            self._add_btn.setFixedWidth(font.BOLD_FONT_METRICS.height() + 8)
+            self._add_btn.setToolTip("Add New User")
+            toolbar.addWidget(self._add_btn)
+
+            # noinspection PyUnresolvedReferences
+            self._add_btn.clicked.connect(self._on_add_btn_clicked)
+
         toolbar.addStretch()
 
-        self._table_view: table_view.TableView[domain.User, str] = table_view.TableView(
-            attrs=(
-                table_view.text(
-                    name="user_id",
-                    display_name="ID",
-                    key=True,
-                ),
-                table_view.text(
-                    name="username",
-                    display_name="Username",
-                    width=200,
-                ),
-                table_view.text(
-                    name="display_name",
-                    display_name="Name",
-                    width=200,
-                ),
-                table_view.timestamp(
-                    name="date_added",
-                    display_name="Added",
-                ),
-                table_view.timestamp(
-                    name="date_updated",
-                    display_name="Updated",
-                    width=100,
-                ),
+        attrs: tuple[table_view.Attr[domain.User, typing.Any], ...] = (
+            table_view.text(
+                name="user_id",
+                display_name="ID",
+                key=True,
+            ),
+            table_view.text(
+                name="username",
+                display_name="Username",
+                width=200,
+            ),
+            table_view.text(
+                name="display_name",
+                display_name="Name",
+                width=200,
+            ),
+            table_view.timestamp(
+                name="date_added",
+                display_name="Added",
+            ),
+            table_view.timestamp(
+                name="date_updated",
+                display_name="Updated",
+                width=100,
+            ),
+        )
+
+        if self._current_user.is_admin:
+            attrs += (
                 table_view.button(
                     name="edit",
                     button_text="",
                     icon=icons.edit_btn_icon(parent=self),
                     width=40,
-                    enabled_when=lambda user: self._enabled_when(user=user),
+                    enabled_when=lambda _: self._current_user.is_admin,
                 ),
                 table_view.button(
                     name="delete",
                     button_text="",
                     icon=icons.delete_btn_icon(parent=self),
                     width=40,
-                    enabled_when=lambda user: self._enabled_when(user=user),
+                    enabled_when=lambda _: self._current_user.is_admin,
                 ),
-            ),
+            )
+
+        self._table_view: table_view.TableView[domain.User, str] = table_view.TableView(
+            attrs=attrs,
             parent=self,
         )
 
@@ -95,11 +107,10 @@ class UserDash(qtw.QWidget):
         self.setLayout(layout)
 
         # noinspection PyUnresolvedReferences
-        self._add_btn.clicked.connect(self._on_add_btn_clicked)
-        # noinspection PyUnresolvedReferences
         self._refresh_btn.clicked.connect(self._on_refresh_btn_clicked)
         self._table_view.button_clicked.connect(self._on_table_btn_clicked)
-        self._table_view.double_click.connect(self._on_table_btn_double_clicked)
+        if self._current_user.is_admin:
+            self._table_view.double_click.connect(self._on_table_btn_double_clicked)
 
     def get_state(self) -> UserDashState:
         return UserDashState(
@@ -122,16 +133,20 @@ class UserDash(qtw.QWidget):
             if state.selected_user is not None:
                 self._table_view.select_item_by_key(state.selected_user.user_id)
 
-    def _enabled_when(self, *, user: domain.User) -> bool:
-        return domain.permissions.user_can_edit_user(
-            current_user=self._current_user,
-            user=user,
-        )
+        if state.user_added:
+            self._table_view.add_item(state.user_added)
+
+        if state.user_deleted:
+            self._table_view.delete_item(key=state.user_deleted.user_id)
+
+        if state.user_updated:
+            self._table_view.update_item(state.user_updated)
 
     def _on_add_btn_clicked(self, /, _: bool) -> None:
         logger.debug(f"{self.__class__.__name__}._on_add_btn_clicked()")
 
-        self._requests.add.emit()
+        if self._current_user.is_admin:
+            self._requests.add.emit()
 
     def _on_refresh_btn_clicked(self, /, _: bool) -> None:
         self._requests.refresh.emit()
@@ -141,20 +156,14 @@ class UserDash(qtw.QWidget):
 
         match event.attr.name:
             case "delete":
-                if domain.permissions.user_can_edit_user(
-                    current_user=self._current_user,
-                    user=event.item,
-                ):
+                if self._current_user.is_admin:
                     if popup.confirm(
                         question=f"Are you sure you want to delete {event.item.display_name}?",
                         title="Confirm Delete",
                     ):
                         self._requests.delete.emit(requests.DeleteRequest(user=event.item))
             case "edit":
-                if domain.permissions.user_can_edit_user(
-                    current_user=self._current_user,
-                    user=event.item,
-                ):
+                if self._current_user.is_admin:
                     self._requests.edit.emit(requests.EditRequest(user=event.item))
             case _:
                 logger.error(
@@ -165,9 +174,6 @@ class UserDash(qtw.QWidget):
     def _on_table_btn_double_clicked(self, /, event: table_view.DoubleClickEvent[domain.User, str]) -> None:
         logger.debug(f"{self.__class__.__name__}._on_table_btn_double_clicked({event=!r})")
 
-        if domain.permissions.user_can_edit_user(
-            current_user=self._current_user,
-            user=event.item,
-        ):
+        if self._current_user.is_admin:
             request = requests.EditRequest(user=event.item)
             self._requests.edit.emit(request)
